third_party/WebKit/LayoutTests/fast/dom/custom/lifecycle-created-paste-expected.txt

This test ensures that the lifecycle callbacks of editing-originated elements are visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.callbacksCalled is ['DIV', 'X-FOO']
PASS window.callbacksCalled is ['DIV', 'X-FOO', 'DIV', 'X-FOO']
PASS successfullyParsed is true

TEST COMPLETE
